requirements.txt
watchdog
websockets
pillow
asyncio